Fixed errors from merge
